perf/requirements.txt
azure-core==1.25.1
azure-storage-blob==12.13.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
idna==3.4
isodate==0.6.1
kafka-python==2.0.2
msrest==0.7.1
oauthlib==3.2.1
prettytable==3.4.1
pycparser==2.21
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
